requirements.txt
requests
colorama
geocode